+++ EndToEndTests/SampleUseTest.php
@@ -32,8 +32,8 @@ class Epa_EndToEndTests_SampleUseTest extends PHPUnit_Framework_TestCase
 
 		# # Plugin style
 		#
-		# The EventDispatcher is a class that is also an observer and thus can
-		# be added like the above `SuccessLoginLogger`. It listens to events
+		# The EventDispatcher is an interface that extends that of an observer and
+		# thus can be added like the above `SuccessLoginLogger`. It listens to events
 		# but dispatches the events to plugins. Plugins register themselves with
 		# the EventDispatcher with `registerPlugin`. The EventDispatcher then calls
 		# them back giving these plugins the chance to register callbacks for certain
@@ -45,7 +45,7 @@ class Epa_EndToEndTests_SampleUseTest extends PHPUnit_Framework_TestCase
 		# Plugins register callbacks to events and this registration is by
 		# the class name of the events (this can be changed, see below).
 
-		$eventDispatcher = new \Epa\EventDispatcher();
+		$eventDispatcher = \Epa\EventDispatcherFactory::create();
 		$login->addObserver($eventDispatcher); // EventDispatcher = observer style
 
 		# The `SuccessLoginLogger` implements the `Plugin` interface and
@@ -57,7 +57,7 @@ class Epa_EndToEndTests_SampleUseTest extends PHPUnit_Framework_TestCase
 		#     );
 
 		$logPlugin = new \Epa\EndToEndTests\Support\SuccessLoginLogger();
-		$eventDispatcher->registerPlugin($logPlugin);
+		$eventDispatcher->addPlugin($logPlugin);
 
 		$login->login('bar', 'baz');
 		$this->assertEquals('success for bar:baz', $logPlugin->getLog());
@@ -67,39 +67,28 @@ class Epa_EndToEndTests_SampleUseTest extends PHPUnit_Framework_TestCase
 		# change the way events are named. We could change this from using
 		# the fully qualified class name to only the classname itself
 		# (eg `Foo` instead of `\MyNamespace\Events\Foo`). That's because the
-		# EventDispatcher throws the `Epa\\NewEventEvent` when it is notified
+		# EventDispatcher throws the `Epa\Api\NewEventEvent` when it is notified
 		# of a new event. This event has the option to add an event name.
 
 		$eventNameChanger = new \Epa\EndToEndTests\Support\EventNameChangerPlugin();
-		$eventDispatcher->registerPlugin($eventNameChanger);
+		$eventDispatcher->addPlugin($eventNameChanger);
 
 		# We'll add a loginLogger that registers a callback for the event `LoginEvent`
 		# instead of the standard `Epa\EndToEndTests\Support\LoginEvent`. The
 		# EventNameChangerPlugin added this version of event names to events.
 
 		$loginEventLogger = new \Epa\EndToEndTests\Support\LoginEventLogger();
-		$eventDispatcher->registerPlugin($loginEventLogger);
+		$eventDispatcher->addPlugin($loginEventLogger);
 
 		# The `LoginEventLogger` wants to be notified when a
 		# `LoginEvent` is thrown. Our 'login' application throws the event
 		# `\Epa\EndToEndTests\Support\LoginEvent` but our `EventNameChangerPlugin`
-		# is notified of the `\Epa\NewEventEvent` and adds a new event name so that
-		# callbacks registered for `LoginEvent` instead of the full class name
-		# also get notified. Let's test it.
+		# is notified of the `\Epa\Api\NewEventEvent` and adds a new event name so that
+		# callbacks registered for `LoginEvent` (instead of the full class name)
+		# also get notified.
 
 		$login->login('foo', 'bar');
 		$this->assertEquals('last login was foo:bar', $loginEventLogger->getLastLogin());
-
-		# A similar plugin already exists (`Epa\MetaEventNamePlugin`). It adds
-		# event names that are specified inthe doc comments of the classes/interfaces
-		# an event extends/implements. `@eventname` adds the interface or class name
-		# as an event. `@eventName someName` adds the specified name as an event.
-		#
-		#     /**
-		#      * @eventName Foo 
-		#      */
-		#
-		# This will add the name 'Foo' as an eventname.
 
 		# When you register a callback for an event the callbacks are notified in
 		# order of registration. Sometimes you may want to add a callback before
@@ -110,8 +99,8 @@ class Epa_EndToEndTests_SampleUseTest extends PHPUnit_Framework_TestCase
 		#       ->first();
 		#
 		# Let's add another plugin that registers a callback for the event
-		# `LoginEvent`, like the `LoginEventLogger` above, but adds it before and
-		# has the callback change the username.
+		# `LoginEvent` but adds the callback before the one `LoginEventLogger`
+		# registers (using `first`). It will change the username used to log in.
 
 		$loginNameChanger = new \Epa\EndToEndTests\Support\LoginNameChanger();
 		$eventDispatcher->registerPlugin($loginNameChanger);