+++ UnitTests/Api/ObserverStoreTest.php
@@ -1,8 +1,10 @@
 <?php
 
-require_once('TestHelper.php');
+require_once dirname(__FILE__)
+		. DIRECTORY_SEPARATOR . '..'
+		. DIRECTORY_SEPARATOR . 'TestHelper.php';
 
-class Epa_UnitTests_PluggableTest extends PHPUnit_Framework_TestCase
+class Epa_UnitTests_Api_ObserverStoreTest extends PHPUnit_Framework_TestCase
 {
 	/**
 	 * @test
@@ -10,8 +12,8 @@ class Epa_UnitTests_PluggableTest extends PHPUnit_Framework_TestCase
 	public function notifiesObserversOfEvent()
 	{
 		$observableClass = new \Epa\UnitTests\Support\ClassWithEvent();
-		$observer = $this->getMock('Epa\\Observer');
-		$event = $this->getMock('Epa\\Event');
+		$observer = $this->getMock('Epa\\Api\\Observer');
+		$event = $this->getMock('Epa\\Api\\Event');
 
 		$observer->expects($this->once())->method('notify')->with($event);